Refactor + publish packet topic alias binding
